+++ CHANGELOG.md
@@ -6,6 +6,19 @@ All notable changes to this project will be documented in this file. Take a look
 
 <!-- ## [Unreleased] -->
 
+## [3.0.1]
+
+### Fixed
+
+#### Shared
+
+* Improved performance when accessing a publication from the Shared Storage.
+
+#### Navigator
+
+* Fixed crash in the image navigator.
+
+
 ## [3.0.0]
 
 ### Fixed
@@ -888,4 +901,5 @@ progression. Now if no reading progression is set, the `effectiveReadingProgress
 [3.0.0-beta.1]: https://github.com/readium/kotlin-toolkit/compare/3.0.0-alpha.2...3.0.0-beta.1
 [3.0.0-beta.2]: https://github.com/readium/kotlin-toolkit/compare/3.0.0-beta.1...3.0.0-beta.2
 [3.0.0]: https://github.com/readium/kotlin-toolkit/compare/3.0.0-beta.2...3.0.0
+[3.0.1]: https://github.com/readium/kotlin-toolkit/compare/3.0.0...3.0.1